% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan-te.R
\name{check_clusters}
\alias{check_clusters}
\title{Check the temporal sequence constraint and return
stop points}
\usage{
check_clusters(trajectory, delta_t, minpts, entr_t)
}
\arguments{
\item{trajectory}{A simple features collection of points representing a
single trajectory (i.e. person, day). MUST have a `dttm` column called
"timestamp" representing the time of GPS point.}

\item{delta_t}{time threshold (seconds): a gap of this length within a
spatial cluster will split the cluster into two potential activities.
An activity must also be at least this long.}

\item{minpts}{minimum number of points in a cluster. DBSCAN input.}

\item{entr_t}{entropy threshold: the entropy of a cluster is a function of
the chaotic movement between points in a cluster. Clusters with higher
entropy are more likely to be activities; this parameter will exclude
potential activities below this threshold}
}
\description{
Check the temporal sequence constraint and return
stop points
}
