% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan-te.R
\name{do_dbscan}
\alias{do_dbscan}
\title{Execute dbscan on trajectory data}
\usage{
do_dbscan(trajectory, eps, minpts)
}
\arguments{
\item{trajectory}{A simple features collection of points representing a
single trajectory (i.e. person, day). MUST have a `dttm` column called
"timestamp" representing the time of GPS point.}

\item{eps}{distance threshold for defining clusters. Should be provided
in the same units as the trajectory projection (usually meters). DBSCAN input.}

\item{minpts}{minimum number of points in a cluster. DBSCAN input.}
}
\description{
Execute dbscan on trajectory data
}
