% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDBC.R
\name{fast_density}
\alias{fast_density}
\title{Fast Density}
\usage{
fast_density(trajectory, d_threshold)
}
\arguments{
\item{trajectory}{A projected sf points dataframe representing a trajectory.}

\item{d_threshold}{A distance (in units of the projection system) to include
in the cluster search.}
}
\description{
The algorithm has its basis in the assumptions that cluster centers are
surrounded by neighbors with lower local density and that they are at a
relatively large distance from any points with a higher local density.
}
\details{
For each data point `i`, we compute two quantities: its
  local density `rho_i` and its distance `delta_i` from points of
  higher density. Both these quantities depend only on the distances Embedded
  Image between data points, which are assumed to satisfy the triangular
  inequality. The local density Embedded Image of data point Embedded Image
  is defined as
}
