% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan-te.R
\name{get_cluster_entropy}
\alias{get_cluster_entropy}
\title{Calculate the entropy of cluster rays}
\usage{
get_cluster_entropy(d)
}
\arguments{
\item{d}{a tibble with a point geometry column}
}
\value{
The cluster entropy
}
\description{
Calculate the entropy of cluster rays
}
\details{
Equation 1 in Gong, et al. (2018)
}
