% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDBC.R
\name{local_density}
\alias{local_density}
\title{Calculate local point density}
\usage{
local_density(d, d_threshold)
}
\arguments{
\item{d}{A vector of distances between a point and all other
points in the trajectory}
}
\description{
Calculate local point density
}
