% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDBC.R
\name{min_distance}
\alias{min_distance}
\title{Find minimum distance to point with lower density}
\usage{
min_distance(d, rho, rho_i)
}
\arguments{
\item{d}{vector of distances}

\item{rho}{vector of point densities}

\item{rho_i}{point density for point i}
}
\description{
Find minimum distance to point with lower density
}
\details{
`delta_i` is measured by computing the minimum distance between the
  point `i` and any other point with higher density:
}
