% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan-te.R
\name{split_cluster}
\alias{split_cluster}
\title{Split clusters based on a delta_t}
\usage{
split_cluster(cluster, delta_t)
}
\description{
Split clusters based on a delta_t
}
\details{
This is a recursive algorithm that divides spatially
  defined clusters based on gaps in their timeline.
}
