% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDBC.R
\name{tdbc}
\alias{tdbc}
\title{Determine stop locations in trajectory data via TDBC}
\usage{
tdbc(trajectory)
}
\arguments{
\item{trajectory}{A simple features collection representing a single
trajectory.}

\item{eps}{distance threshold for defining clusters}

\item{minpts}{minimum number of points in a cluster}
}
\description{
Determine stop locations in trajectory data via TDBC
}
\details{
Implements the method describes in Fu, Z., Tian, Z., Xu, Y., &#38;
   Qiao, C. (2016). A Two-Step Clustering Approach to Extract Locations from
   Individual GPS Trajectory Data. ISPRS International Journal of
   Geo-Information 2016, Vol. 5, Page 166, 5(10), 166.
   \url{https://doi.org/10.3390/IJGI5100166}
}
\examples{
trajectory <- trajectories \%>\%  filter(id == "131") \%>\%
  mutate(day = day(localtime)) \%>\% filter(day == 16)

}
